Fix "array subscript has type `char'" warning.
Cygwin libc is newlib, not glibc.
Newlib implements isalpha(), toupper() and tolower() as macro.
So, do not use them with raw argument typed `char' because it may be
signed.
Also, do not increment/decrement the argument at the same time to avoid
unexpected side effect.

+++ src/type.c
@@ -752,7 +752,7 @@ const char *type_pp(type_t t)
          char *p = fn;
          const char *end = fn + sizeof(fn);
          const char *fname = istr(type_ident(t));
-         const bool operator = !isalpha(fname[0]);
+         const bool operator = !isalpha((uint8_t)fname[0]);
          const char *quote = operator ? "\"" : "";
 
          p += snprintf(p, end - p, "%s%s%s(", quote, fname, quote);
